fix: 修复 sumOf 类型错误 (Float 转 Double)

+++ app/src/main/java/com/gymlog/app/data/repository/WorkoutRepository.kt
@@ -156,7 +156,7 @@ class WorkoutRepository(private val dao: WorkoutDao) {
             
             val totalVolume = completedExercise.sets
                 .filter { it.completed }
-                .sumOf { it.weight * it.reps }
+                .sumOf { (it.weight * it.reps).toDouble() }
             
             val existingPB = dao.getPersonalBestByExercise(completedExercise.exerciseId)
